refactor: 拡張しやすいように
